feat-bo: extract table logic from table components to TableWrapper component and add it to Admins page
